refactor(auth): handler -> interfaces로 구조 변경
