+++ src/main/kotlin/com/groom/auth/domain/authentication/AuthenticationRepository.kt
@@ -1,6 +1,6 @@
 package com.groom.auth.domain.authentication
 
-interface AuthenticationRepository {
+internal interface AuthenticationRepository {
     fun create(initialRole: AuthenticationRole = AuthenticationRole.USER): Authentication
     fun findBy(id: Long): Authentication
 }
