+++ src/main/kotlin/com/groom/auth/domain/authentication/AuthenticationService.kt
@@ -3,7 +3,7 @@ package com.groom.auth.domain.authentication
 import org.springframework.stereotype.Component
 
 @Component
-class AuthenticationService(private val authenticationRepository: AuthenticationRepository) {
+class AuthenticationService internal constructor(private val authenticationRepository: AuthenticationRepository) {
     fun read(id: Long): Authentication = authenticationRepository.findBy(id)
     fun create(initializeRole: AuthenticationRole = AuthenticationRole.USER): Authentication {
         return authenticationRepository.create(initializeRole)
