+++ src/main/kotlin/com/groom/auth/domain/oauth2/OAuth2UserInformationRepository.kt
@@ -1,6 +1,6 @@
 package com.groom.auth.domain.oauth2
 
-interface OAuth2UserInformationRepository{
+internal interface OAuth2UserInformationRepository{
     fun findBy(providerName: OAuth2ProviderName,
                providerUserId: String): OAuth2UserInformation?
