+++ src/main/kotlin/com/groom/auth/domain/oauth2/OAuth2UserInformationService.kt
@@ -3,7 +3,7 @@ package com.groom.auth.domain.oauth2
 import org.springframework.stereotype.Service
 
 @Service
-class OAuth2UserInformationService(private val oAuth2UserInformationRepository: OAuth2UserInformationRepository) {
+class OAuth2UserInformationService internal constructor(private val oAuth2UserInformationRepository: OAuth2UserInformationRepository) {
     fun read(providerName: OAuth2ProviderName, providerUserId: String): OAuth2UserInformation? =
         oAuth2UserInformationRepository.findBy(providerName, providerUserId)
