+++ src/main/kotlin/com/groom/auth/interfaces/api/AuthController.kt
@@ -1,6 +1,5 @@
-package com.groom.auth.interfaces
+package com.groom.auth.interfaces.api
 
-import com.groom.auth.CustomOAuth2User
 import com.groom.auth.component.JwtGenerator
 import com.groom.common.Response
 import org.springframework.security.oauth2.client.registration.ClientRegistrationRepository
@@ -20,7 +19,7 @@ class AuthController(private val oAuth2UserService: OAuth2UserService<OAuth2User
                      private val clientRegistrationRepository: ClientRegistrationRepository,
                      private val jwtGenerator: JwtGenerator) {
     @PostMapping("/login/oauth2/token/{providerName}")
-    fun oAuth2LoginWithToken(@PathVariable providerName: String,
+    internal fun oAuth2LoginWithToken(@PathVariable providerName: String,
                              @RequestBody body: AuthRequest.OAuth2LoginWithAccessToken): Response<AuthResponse.Login> {
         val clientRegistration = clientRegistrationRepository.findByRegistrationId(providerName)
         val request = OAuth2UserRequest(clientRegistration, body.token)
