+++ src/main/kotlin/com/groom/auth/interfaces/api/AuthResponse.kt
@@ -1,7 +1,7 @@
-package com.groom.auth.interfaces
+package com.groom.auth.interfaces.api
 
 import org.springframework.security.oauth2.core.OAuth2AccessToken
 
-class AuthResponse private constructor() {
+internal class AuthResponse private constructor() {
     data class Login(val accessToken: OAuth2AccessToken)
 }
